backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Werkzeug==2.3.7
SQLAlchemy==1.4.23
python-dotenv==0.19.0
psycopg2-binary==2.9.1
gunicorn==21.2.0
requests==2.26.0
bcrypt==4.1.2
cloudinary==1.33.0 